Provide bindings for Vala
Bindings are installed in default location '/usr/share/vala/vapi' along
with the dependencies.

The 'msgpack_object_union' is separated in multiple simple types that
inherit from 'MessagePack.Union'.

APIs using 'long long' are stubbed with 'long' since Vala does not
understand that datatype. It is implicitly converted in the generated C
code.